Implementar la clase PasswordHasher en el mismo archivo de la clase SingleResponsabilityExample
