+++ src/main/java/org/example/singleResponsibility/SingleResponsabilityExample.java
@@ -14,3 +14,20 @@ public class SingleResponsabilityExample {
     }
 }
 
+class PasswordHasher {
+    public String hashAndSavePassword(String password) {
+        hashPassword();
+        savePassword();
+        return "Password hashed and saved";
+    }
+
+    public void hashPassword() {
+        //hashing implementation
+    }
+
+    public void savePassword() {
+        //save to the db
+    }
+}
+
+
